==> 05_Add_MILKYS_data.Rmd <==
---
title: "05. Add Milkys metadata"
author: "DHJ"
date: "2 7 2019"
output: 
  html_document:
    keep_md: true
    toc: true
    toc_float: true
    number_sections: true
    
---

In this script, we base ourself on the output of script 04. We then use the data from the MILKYS Access database to add columns for dry weight, fat weight and fish length into the data (which is also here called dat2, as in scroipt 04, but it is not the same dat2...). We do this by simply deleting all Milkys cod (there is also Urban fjord cod) and put in new cod data made from the Access base. As in script 04, we also make "wide-form data" (version 3 of the Excel files named 002,003,004) at the end of the script. 

## Setup

### Libraries
```{r}
library(dplyr)
library(niRvana)
library(ggplot2)
library(tidyr)

# install.packages("ODBC")
source("05_Add_MILKYS_data_functions.R")
```

### Set username and password
```{r}
# set_credentials()
```

## Data  
### Read data  
- dat: Excel data from Merete
- dat2_means: based on Excel data, species and tissua added, no less-thans (not used here)
- cemp_comb: Access data 
- data_ind2: Access data, updated with NIVAbasen data through 2017
```{r}
# Data from Merete (stacked data)
dat <- readxl::read_excel("Input_data/001 PFAS stacked data table.xlsx", guess_max = 11000)

# From script 05 - 'dat2_means' (duplicate-free version of 'dat2')
dat2 <- readRDS(file = "Data/04_dat2.rds")
dat2_means <- readRDS(file = "Data/04_dat2_means.rds")

# Access data
load("Input_data/Milkys_Data_2017_02_14_cemp_combinedtable.RData")  # cemp_comb
# load("Input_data/Milkys_Data_2017_02_14_cemp_tables.RData")  # cemp_c2, cemp_f2, cemp_t 

# Access data with NIVAbasen data added for 2015-2017
fn <- "../Milkys_2018/Data/12_data_ind2_2018-08-23.RData"
data_ind2 <- readRDS(fn)
```


### cemp_comb + data_ind2: Add 'param2'
```{r}
cemp_comb <- cemp_comb %>%
  left_join(cemp_synonyms) %>%
  mutate(param2 = ifelse(is.na(param_standard), param, param_standard)) %>%
  select(-param_standard)

cemp_comb %>%
  filter(param != param2) %>%
  count(param, param2)

data_ind2 <- data_ind2 %>%
  left_join(cemp_synonyms, by = c("PARAM" = "param")) %>%
  mutate(param2 = ifelse(is.na(param_standard), PARAM, param_standard)) %>%
  select(-param_standard)
```


### Same station/year in 4 datasets  
```{r}
# Access database
test_cemp <- cemp_comb %>%
  filter(jmpst %in% "36B" & myear %in% 2013 & param %in% "PFOS") 
# View(test_cemp)
test_cemp

# Access + NIVAbasen database
test_cemp2 <- data_ind2 %>%
  filter(STATION_CODE %in% "36B" & MYEAR %in% 2013 & param2 %in% "PFOS") 
# View(test_cemp2)
test_cemp2

#
# Commented out in order to 
#
# Database (WILAB)
# test_wilab <- get_nivabase_data("select * from WILAB.TEST where TESTNO like '2013-02901'")
# test_wilab %>% select(TESTNO, SERIALNO, MARKING, DESCRIPT, X5, X9)

# Excel data
sel <- with(dat, substr(LIMS,4,7) == "2013" & Species %in% "Gadus morhua" & Description %in% "36B" & PFAS %in% "PFOS")
sum(sel)
# View(dat[sel, ])
dat[sel,] %>% select(Description, Project, LIMS, Matrix_orig, Matrix, Species, Organ_orig, Organ, Label_original, PFAS, Data)
```



### Number of PFAS params  
For the example station/year
```{r}
cat("Excel data:\n")
dat %>%
  filter(substr(LIMS,4,7) == "2013" & Species %in% "Gadus morhua" & Description %in% "36B") %>%
  pull(PFAS) %>%
  unique() %>% sort()

# Access
df <- cemp_comb %>%
  filter(jmpst %in% "36B" & myear %in% 2013 & substr(param,1,2) == "PF")
cat("\nAccess, new names:\n")
df %>%
  pull(param2) %>%
  unique() %>% sort()
cat("\nAccess, old names:\n")
df %>%
  pull(param) %>%
  unique() %>% sort()

# Access + NIVAbasen
df <- data_ind2 %>%
  filter(STATION_CODE %in% "36B" & MYEAR %in% 2013 & substr(param2,1,2) %in% "PF") 
cat("\nAccess + NIVAbasen, new names:\n")
df %>%
  pull(param2) %>%
  unique() %>% sort()
cat("\nAccess + NIVAbasen, old names:\n")
df %>%
  pull(PARAM) %>%
  unique() %>% sort()

```


### Number of PFAS params 2
For a different station/year (Sørfjorden)
```{r}
cat("Excel data:\n")
dat %>%
  filter(substr(LIMS,4,7) == "2009" & Species %in% "Gadus morhua" & Description %in% "23B") %>%
  pull(PFAS) %>%
  unique() %>% sort()

# Access
df <- cemp_comb %>%
  filter(jmpst %in% "23B" & myear %in% 2009 & substr(param,1,2) == "PF")
cat("\nAccess, new names:\n")
df %>%
  pull(param2) %>%
  unique() %>% sort()
cat("\nAccess, old names:\n")
df %>%
  pull(param) %>%
  unique() %>% sort()

```

### Recreate cemp_synonyms and use it   
Used only while refining cemp_synonyms (in the ...functions file)
```{r}
# source("05_Add_MILKYS_data_functions.R")
# load("Input_data/Milkys_Data_2017_02_14_cemp_combinedtable.RData")  # cemp_comb
# cemp_comb <- cemp_comb %>%
#   left_join(cemp_synonyms) %>%
#   mutate(param2 = ifelse(is.na(param_standard), param, param_standard)) %>%
#   select(-param_standard)

```
## 11 most common parameters in cod
### Excel data
```{r}
# cat("Excel data:\n")
df <- dat %>%
  filter(Year <= 2014 & Species %in% "Gadus morhua") %>%
  count(PFAS) %>%
  filter(n > 10) %>%
  arrange(PFAS)
df %>% pull(PFAS)
# "PFBS"   "PFDA"   "PFDS"   "PFHpA"  "PFHxA"  "PFHxS"  "PFNA"   "PFOA"   "PFOS"   "PFOSA"  "PFUnDA" - the 11 most common ones
# And the same as in CEMP
df
```

### Access
```{r}
df <- cemp_comb %>%
  filter(myear <= 2014 & grepl("^PF", param2) & speci == "GADU MOR") %>%
  count(param2) %>%
  filter(n > 50) %>%
  arrange(param2)
df %>% pull(param2)
df
# "PFBS"   "PFDA"   "PFDS"   "PFHpA"  "PFHxA"  "PFHxS"  "PFNA"   "PFOA"   "PFOS"   "PFOSA"  "PFUnDA" - the 11 most common ones
```



## The less used parameters in cod
### Excel data
```{r}
# cat("Excel data:\n")
df <- dat %>%
  filter(Year <= 2014 & Species %in% "Gadus morhua") %>%
  count(PFAS) %>%
  filter(n < 10) %>%
  arrange(PFAS)
df %>% pull(PFAS)
# "10:2 FTOH" "12:2 FTOH" "4:2 FTOH"  "6:2 FTOH"  "6:2 FTS"   "8:2 FTOH"  "8:2 FTS"   "N-EtFOSA"  "N-EtFOSE"  "N-MeFOSA" 
# "N-MeFOSE"  "PFBA"      "PFDoDA"    "PFDoS"     "PFHpS"     "PFNS"      "PFPA"      "PFPeS"     "PFPS"      "PFTeDA"   
# "PFTeS"     "PFTrDA"    "PFTrS"     "PFUnS"
df
```


### Access
```{r}
df <- cemp_comb %>%
  filter(myear <= 2014 & grepl("^PF", param2) & speci == "GADU MOR") %>%
  count(param2) %>%
  filter(n <= 50) %>%
  arrange(param2)
df %>% pull(param2)
# Here:            "PFBA"   "PFDoDA" "PFHXDA" "PFODA"  "PFPeA"  "PFTeDA" "PFTrDA"
# In Excel data:     x         x                                   x        x
df
```

## Number of observations, cod (PFOS obs per year)
### Excel data
```{r}
# cat("Excel data:\n")
dat %>%
  filter(Year <= 2014 & Species %in% "Gadus morhua" & PFAS == "PFOS" & Description == "23B") %>%
  count(Year)
```

### Access  
The number of observations in Access is much higher and makes much more sense
```{r}
cemp_comb %>%
  filter(myear <= 2014 & speci == "GADU MOR" & param2 %in% "PFOS" & jmpst %in% "23B") %>%
  count(myear)
```

## Stations, cod (PFOS as example)
### Excel data
```{r}
# cat("Excel data:\n")
dat %>%
  filter(Year <= 2014 & Species %in% "Gadus morhua" & PFAS == "PFOS") %>%
  count(Description)
```

### Access  
The same stations, but lacking '13BH' and with 145B, 146B and 147B in addition
```{r}
cemp_comb %>%
  filter(myear <= 2014 & speci == "GADU MOR" & param2 %in% "PFOS") %>%
  count(jmpst)
```

## Number of observations, blue mussel (PFOS obs per year)
### Excel data  
Extremely few observations up until 2014 (and not so many thereafter)  
```{r}
# dat %>%
#   count(Species)
dat %>%
  filter(Year <= 2014 & Species %in% "Blåskjell" & PFAS == "PFOS") %>%
  count(Description, Year)
```

### Access  
Even fewer observations  
```{r}
cemp_comb %>%
  filter(speci == "MYTI EDU" & param2 %in% "PFOS") %>%
  count(speci, jmpst, myear)
# cat("Excel data:\n")
```
## Make ready for making new version of data set  
- Delete MilKys cod through 2017, and replace those data with cod data from Access   
- Only 
- Also add new columns for fat, dry weight,length, weight   

### Counts
```{r}
data_ind2 %>% filter(LATIN_NAME %in% "Gadus morhua" & param2 %in% "PFOS") %>% nrow()
dat2_means %>% filter(Species %in% "Gadus morhua" & PFAS %in% "PFOS") %>% nrow()
```

### Tissues
Lever only  
```{r}
dat2 %>% 
  filter(Project %in% "MilKys") %>%
  count(Project, Species, Organ)

data_ind2 %>% 
  filter(LATIN_NAME %in% c("Gadus morhua", "Platichthys flesus", "Limanda limanda")) %>%
  filter(grepl("^PF", PARAM)) %>%
  count(LATIN_NAME, TISSUE_NAME) # %>% pull(TISSUE_NAME)
```

### Units 1
Access data = U, NIVAbasen = UG_P_KG, Excel = ng/g  
We test below that they are the same  
```{r}
data_ind2 %>%
  filter(LATIN_NAME %in% "Gadus morhua" & param2 %in% "PFOS") %>%
  xtabs(~MYEAR + UNIT, .)

dat %>%
  filter(Species %in% "Gadus morhua" & PFAS %in% "PFOS") %>%
  count(Unit)
```
### Units 2 - test using plot (2013)  
Unit UG_P_KG in Access + NIVAbase, vs ng/g for Excel  
```{r}
# Access + NIVAbasen database
library(cowplot)

test1 <- dat %>%
  filter(Year == "2013" & Species %in% "Gadus morhua" & Description %in% "36B" & PFAS %in% "PFOS") %>%
  mutate(Value = as.numeric(sub("<","", Data)),
         Flag = ifelse(grepl("<", Data), "<", as.character(NA)))
gg1 <- ggplot(test1, aes(x = Flag, y = Value)) + 
  geom_point() +
  labs(title = "Access + NIVAbase (2013)")

# par(mfrow = c(1,2), mar = c(4,5,2,1))
# plot(test1$Value)

# Excel data

test2 <- data_ind2 %>%
  filter(STATION_CODE %in% "36B" & MYEAR %in% 2013 & param2 %in% "PFOS") 
gg2 <- ggplot(test2, aes(x = FLAG1, y = VALUE_WW)) + 
  geom_point() + 
  labs(title = "Excel (2013)")

plot_grid(gg1, gg2, ncol = 2)
```

### Units 3 - test using plot (2009)
Unit U in Access + NIVAbase, vs ng/g for Excel  
```{r}
# Access + NIVAbasen database
library(cowplot)

test1 <- dat %>%
  filter(Year == "2009" & Species %in% "Gadus morhua" & Description %in% "36B" & PFAS %in% "PFOS") %>%
  mutate(Value = as.numeric(sub("<","", Data)),
         Flag = ifelse(grepl("<", Data), "<", as.character(NA)))
gg1 <- ggplot(test1, aes(x = Flag, y = Value)) + 
  geom_point() +
  labs(title = "Access + NIVAbase (2009)")

# par(mfrow = c(1,2), mar = c(4,5,2,1))
# plot(test1$Value)

# Excel data

test2 <- data_ind2 %>%
  filter(STATION_CODE %in% "36B" & MYEAR %in% 2009 & param2 %in% "PFOS") 
gg2 <- ggplot(test2, aes(x = FLAG1, y = VALUE_WW)) + 
  geom_point() + 
  labs(title = "Excel (2009)")

plot_grid(gg1, gg2, ncol = 2)
```
## Make new version of data set (dat2)

```{r}
sel <- is.na(dat$Data)
mean(sel)  # 0.00087
```

### Make data set 'dat_withoutcod' by deleting the cod data  
And also adding Value amd Flag as well as dryweight, fat percent and 
```{r}
# 
dat_withoutcod <- dat %>%
  filter(!(Species %in% "Gadus morhua" & Project %in% "MilKys")) %>%  # Remove MilKys fish
  mutate(Value = as.numeric(sub("<","", Data)),                     # Make Value from Data
         Flag = ifelse(grepl("<", Data), "<", as.character(NA)),    # Make Flag (less-than sign)
         Drywt = as.numeric(NA),
         Fatperc = as.numeric(NA),
         Length = as.numeric(NA))

nrow(dat)
nrow(dat_withoutcod)
```

### Make data for adding ('data_to_add')
Note that we "construct" new values for "LIMS" on a completely different format, as this is needed to 
identify individuals. These new identifiers (starting with "Nivabase_") have no connection to LIMS.  
```{r}
check <- data_ind2 %>%
  filter(param2 == "PFAS") %>%
  count(STATION_CODE, MYEAR, SAMPLE_NO2)
table(check$n)
check %>% filter(n == 2)   # Ærfugl has "duplicates" as thy have 2 tissues (eggs + blood) for each station
```


```{r}
params <- c(
  "PFBS", "PFDA", "PFDS", "PFHpA", "PFHxA", "PFHxS", "PFNA", "PFOA", "PFOS", "PFOSA", "PFUnDA",
  "PFBA", "PFDoDA", "PFHXDA", "PFODA", "PFPeA", "PFTeDA", "PFTrDA"
  )

data_to_add <- data_ind2 %>%
  filter(LATIN_NAME %in% "Gadus morhua") %>%
  filter(param2 %in% params) %>%
  rename(
    Year = MYEAR,
    Description = STATION_CODE,
    Species = LATIN_NAME,
    Organ = TISSUE_NAME,
    PFAS = param2,
    Value = VALUE_WW,
    Flag = FLAG1,
    Drywt = DRYWT,
    Fatperc = FAT_PERC,
    Length = LNMEA
  ) %>%
  mutate(
    Organ = "liver",    # In this special case (see part named "Tissues")
    Unit = "ng/g",
    LIMS = paste0("Access_", Description, Year, SAMPLE_NO2),   # Note: construct new values for 'LIMS'
    Matrix = "biota",
    Class = "Fish") %>%
  select(-c(SAMPLE_NO2, BASIS, UNIT, PARAM,
            VALUE_WWa, VALUE_DW, VALUE_DWa, VALUE_FB, VALUE_FBa))
```
[1] NA           "blood"      "egg"        "filet"      "liver"      "whole body"
[1] "Blod"   "Galle"  "Lever"  "Muskel"

### Add PFASgroup etc
```{r}
# Make data frame for joining to data_to_add
dat_pfasgroups <- dat %>%
  group_by(PFAS) %>%
  summarise_at(c("PFASgroup","PFASlength","PFASlength_nr"), first)

data_to_add <- left_join(data_to_add, dat_pfasgroups)
```
### Combine data  
```{r}
dat2 <- bind_rows(dat_withoutcod, data_to_add)
```

### PFPA vs PFPeA: Show difference (different Label_original, lumped togehther in same Label)
```{r}
# 1. Some Milkys PFPA data 2006-2007 - two different 'Label_original': PFPA and PFPeA
# We take care of this below

# Example
dat2 %>% 
  filter(LIMS %in% "NR-2007-60795" & PFAS == "PFPA") %>% 
  select(LIMS, PFAS, Label_original, Description, Species, Class, Data)
```

### PFPA vs PFPeA: change 'Label'  
```{r}
dat2 <- dat2 %>%
  mutate(PFAS = ifelse(Label_original %in% c("PFPeA", "HPFHpA"), Label_original, PFAS))
```

## Save the new data 'dat2' set
```{r}
# Save in Rdata format
saveRDS(dat2, file = "Data/05_dat2.rds")
# For loading:
# dat2 <- readRDS(file = "Data/05_dat2.rds")

# Save in excel format
infostring <- c(
  "Based on '001 PFAS stacked data table.xlsx'. Cod data replaced with cod data from Access",
  "and NIVAbasen data. This both adds more datapoints (especially for older data) and",
  "new columns Drywt, Fatperc and Length.",
  "Note that we 'construct' new values for LIMS on a completely different format, as this is needed", 
  "to identify individuals. These new identifiers (starting with 'Nivabase_') have no connection", 
  "to LIMS.",
  "Code: 05_Add_MILKYS_data.Rmd"
)
  
data_for_excel <- list(
  dat2,
  tibble(Info = infostring)
)
names(data_for_excel) <- c("Data", "Info")

openxlsx::write.xlsx(data_for_excel, file = "Data/PFAS stacked data table ver 03 (script 05).xlsx")

```


## Check of data
### Test plot
```{r}
dat2 %>% 
  filter(PFAS %in% "PFOS" & is.na(Flag) & Species %in% "Gadus morhua") %>%
  ggplot(aes(Year, Value, color = Organ)) +
  geom_point() +
  facet_wrap(vars(Description)) +
  scale_y_log10()
```

### Test plot for a single parameter     
```{r}
dat2 %>% 
  filter(PFAS %in% "PFOS" & is.na(Flag) & Species %in% "Gadus morhua" & Description %in% "36B") %>%
  ggplot(aes(Year, Value)) +
  geom_jitter(width = 0.15) +
  scale_y_log10()
```

### Test plot PFOSA vs PFOS
```{r}
# Must rearrange data using tidyr::spread (parameters side-by-side)
df_test <- dat2 %>% 
  filter(Species %in% "Gadus morhua" & Description %in% "36B") %>%
  select(Species, Description, Year, LIMS, Length, PFAS, Value) %>%
  tidyr::spread(key = PFAS, value = Value) 

ggplot(df_test, aes(PFOS, PFOSA, color = Year)) +
  geom_point() +
  scale_color_gradient(low = "red", high = "blue") + 
  scale_y_log10() + scale_x_log10()
```



## Prepare for making wide format  
Wide format: One column = one parameter  
Needs to have only one row for each combinations of LIMS, PFAS, Species, etc. That is what we do below.  
  
This part and next part ('Make wide format data') are based on the corresponding parts in script 04, but we use Value instead of Data, as 'Data' is lacking for all cod. Another difference is that PFPA vs PFPeA has already been dealt with above.    

### Same LIMS can have two different Matrix  
Example below. So we need to group on Matrix as well. 
```{r}
dat2 %>% filter(LIMS %in% "NR-2013-2612-1A" & PFAS %in% "PFBS") %>% select(LIMS, Matrix, PFAS, Data)
```

### Check duplicates 1   
ca 783 duplicates, but as we will see in the next part 
```{r}
#  Check whether we have duplicates
check1 <- dat2 %>%
  count(LIMS, PFAS, Matrix) %>%
  mutate(LIMS_PFAS_Matrix = paste(LIMS, PFAS, Matrix, sep = "_"))

table(check1$n)
```

### Check duplicates 2   
check2: pick duplicates from check1 and list all different values for each variable  
Exploration of the 4 different "cases" in check2  
```{r}
# Function for listing all different variables
vals <- function(x) {
  paste(unique(x), collapse = ";")
}
# vals(c(1,4,4,6,8,19,8))

check2 <- dat2 %>%
  mutate(LIMS_PFAS_Matrix = paste(LIMS, PFAS, Matrix, sep = "_")) %>%
  filter(LIMS_PFAS_Matrix %in% subset(check1, n > 1)$LIMS_PFAS_Matrix) %>%
  select(LIMS, Description, Project, Species, Unit, Count, Year, Matrix, Organ, Class,
         PFAS, Value) %>%
  group_by(LIMS, PFAS) %>%
  summarise_all(vals)

sel <- grepl(";", check2$Value); sum(sel)  # 47, so most of the 773 duplicates have the same value
mean(sel)  # 10%

# Look at the whole thing (code below is based on this)
# View(check2[sel,])
```

#### Exploration of each "case":
```{r}
# 1. Store innsjøer 2015 - registered  twice with slightly different concentrations
#    Sometimes also as species Osnerus eperlanus and Krøkle (same thing)
check2[sel,] %>% 
  filter(Project == "Store innsjøer") %>% 
  select(LIMS, PFAS, Description, Species, Class, Value)
# Used later
lims_storeinnsjo_2015 <- check2[sel,] %>% 
  filter(Project == "Store innsjøer") %>% 
  select(LIMS, PFAS, Description, Species, Class, Value) %>%
  pull(LIMS) %>% unique()
dat2 %>% 
  filter(LIMS %in% "NR-2015-09882" & PFAS == "PFDA") %>% 
  select(LIMS, PFAS, Label_original, Description, Species, Class, Value)
# Fish vs Ørret
dat2 %>% 
  filter(LIMS %in% "NR-2015-09889" & PFAS == "PFNA") %>% 
  select(LIMS, PFAS, Label_original, Description, Species, Class, Value)

```


```{r}
# 2. Tyrifjorden 2018 - some with LIMS number "NR-2018-" (but Description differs)
check2[sel,] %>% 
  filter(Project == "Tyrifjorden") %>% 
  select(LIMS, PFAS, Description, Species, Class, Value)
dat2 %>% 
  filter(LIMS %in% "NR-2018-" & PFAS == "PFDA") %>% 
  select(LIMS, PFAS, Label_original, Description, Species, Class, Value)
```

```{r}
# 3. Randsfjorden (I964, I965, I969) 2018 - duplicates of Sum
#    May be with and without detection limit
# Note: in script 04 there were also confusion PFPA and PFPeA, plus PFHpA and HPFHpA  
# The least common ones are not included here in script 05

check2[sel,] %>% 
  filter(Project == "MilKys" & Description %in% c("I964","I965","I969")) %>% 
  select(LIMS, PFAS, Description, Species, Class, Value)
#    Sum: may be with and without detection limit?
dat2 %>% 
  filter(LIMS %in% "NR-2018-242373") %>% 
  select(LIMS, PFAS, Label_original, Description, Value)
# Used later
lims_ranfjorden <- dat2 %>% 
  filter(Project == "MilKys" & Description %in% c("I964","I965","I969") & PFAS %in% "Sum") %>%
  select(LIMS, PFAS, Label_original, Description, Species, Class, Value) %>%
  pull(LIMS) %>% unique()
```

### Make dat2_corrected  
Note: we use only values over detection limit (deleted less-thans)  
```{r}
# For broad data
dat2_corrected <- dat2 %>%
  filter(is.na(Flag)) %>%        # delete less-thans
  mutate(Value = as.numeric(Value)) %>%
  select(LIMS, Label_original, Description, Project, Species, Unit, Count, Year, Matrix, Organ, Class,
         Drywt, Fatperc, Length,
         PFAS, Value)
nrow(dat2_corrected)

# Part 4 above: Delete high sums for Ranfjorden blue mussel 
dat2_corrected <- dat2_corrected %>%
  filter(!(LIMS %in% lims_ranfjorden & PFAS %in% "Sum" & Value > 10))
nrow(dat2_corrected)

# Part 2 above - set same Species
# That values differ slighlty we fix by making the means
dat2_corrected <- dat2_corrected %>%
  arrange(LIMS, PFAS, Species) %>%
  group_by(LIMS, PFAS) %>%
  mutate(Species = ifelse(LIMS %in% lims_storeinnsjo_2015, first(Species), Species)) %>%
  ungroup()
nrow(dat2_corrected) # 23880

# Part 3 should be no problem, since "Description" differs
```

```{r}
sel <- is.na(dat2$Value)
mean(sel)  # 0.00076
# head(dat2[sel,], 1000)

mean(is.na(dat2$Value))
mean(is.na(dat2_corrected$Value))
```

### Make means of the modified data (dat2_means)
#### Preliminary data set dat2_means  
Checks variation of Value using the coeff. of variation (CV)   
```{r}
dat2_means <- dat2_corrected %>%
  filter(!is.na(dat2_corrected$Value)) %>%
  group_by(LIMS, Label_original, Description, Project, Species, Unit, Count, Year, Matrix, Organ, Class,
           Drywt, Fatperc, Length,
           PFAS) %>%
  summarize(Value_mean = mean(Value), Value_sd = sd(Value)) %>%
  mutate(Value_CV = Value_sd/Value_mean)  %>%   # coeff. of variation
  ungroup()

cat("Data size:\n")
nrow(dat2)
nrow(dat2_means)

cat("Coeff. of variation summary:\n")
sel <- is.finite(dat2_means$Value_CV)
sum(sel)  # 24
mean(sel)
summary(dat2_means$Value_CV[sel])

dat2_means[sel,] %>% 
  arrange(desc(Value_CV))

# Check the two highest   
# Accepted for now
lims_highcv <- dat2_means[sel,] %>% filter(Value_CV > 0.1) %>% pull(LIMS)
param_highcv <- dat2_means[sel,] %>% filter(Value_CV > 0.1) %>% pull(PFAS)

dat2 %>% 
  filter(LIMS %in% lims_highcv & PFAS %in% param_highcv) %>% 
  arrange(LIMS, PFAS)
```

#### Accept Value_mean and finish dat2_means   
```{r}
dat2_means <- dat2_means %>%
  rename(Value = Value_mean) %>%
  select(-Value_sd, -Value_CV)
```

### Save dat2_means  
```{r}
saveRDS(dat2_means, "Data/05_dat2_means.rds")
```

### Which PFAS has highest percentage detection?
```{r}
# First, make  
detection_rate <- dat2_corrected %>%
  count(PFAS) %>%
  arrange(desc(n))

# Will be used for factor() later
PFAS_ordered <- detection_rate %>% pull(PFAS)

detection_rate
```

## Make wide format data

### All data
```{r}
# Restructure datausing tidyr::spread
dat2_wide <- dat2_means %>%
  mutate(PFAS = factor(PFAS, levels = PFAS_ordered)) %>%
  select(LIMS, Description, Project, Species, Unit, Count, Year, Matrix, Organ, Class,
         Drywt, Fatperc, Length,
         PFAS, Value) %>%
  tidyr::spread(key = PFAS, value = Value)

# Check number of PFAS per line
df <- dat2_wide %>% 
  select(-(LIMS:Length))
table(apply(!is.na(df), 1, sum))
# apply(!is.na(df), 2, sum)

# Save in Rdata format
saveRDS(dat2_wide, file = "Data/05_dat2_wide.rds")

openxlsx::write.xlsx(dat2_wide, file = "Data/002 PFAS split table ver03.xlsx")
```


### Biota data
```{r}
sum(is.na(dat2_wide$Matrix)) # 0
xtabs(~addNA(Species) + Matrix, dat2_wide)

openxlsx::write.xlsx(dat2_wide %>% filter(Matrix %in% "biota"),
                     file = "Data/003 Subset PFAS biota ver03.xlsx")

```

### Fish data
```{r}
sum(is.na(dat2_wide$Class)) # 2
xtabs(~Species + addNA(Class), subset(dat2_wide, Class %in% "Fish"))

openxlsx::write.xlsx(dat2_wide %>% filter(Class %in% "Fish"),
                     file = "Data/004 Subset PFAS fish ver03.xlsx")

```
